requirements.txt
scikit-learn
numpy
pandas
uvicorn
fastapi
pydantic
setuptools
python-dotenv
pytest